+++ modules/service/service_update_bind.go
@@ -11,7 +11,7 @@ type (
 		Provider    string       `bson:"provider" yaml:"provider"`
 		Deposit     models.Coins `bson:"deposit" yaml:"deposit"`
 		Pricing     string       `bson:"pricing" yaml:"pricing"`
-		QoS         uint64       `bson:"qos" yaml:"qos"`
+		QoS         int64        `bson:"qos" yaml:"qos"`
 		Owner       string       `bson:"owner" yaml:"owner"`
 	}
 )
@@ -27,7 +27,7 @@ func (m *DocMsgUpdateServiceBinding) BuildMsg(v interface{}) {
 	m.Provider = msg.Provider
 	m.Deposit = models.BuildDocCoins(msg.Deposit)
 	m.Pricing = msg.Pricing
-	m.QoS = msg.QoS
+	m.QoS = int64(msg.QoS)
 	m.Owner = msg.Owner
 }
